fixed json, read from json, fixed copy

+++ src/myMath/GuiParams.java
@@ -2,11 +2,11 @@ package myMath;
 
 public class GuiParams 
 {
-	private double width,hight,resolution;
+	private int width,hight,resolution;
 	private Range range_x;
 	private Range range_y;
 	
-	public GuiParams(double width,double hight,double resolution,Range x,Range y)
+	public GuiParams(int width,int hight,int resolution,Range x,Range y)
 	{
 		this.width = width;
 		this.hight = hight;
@@ -19,22 +19,22 @@ public class GuiParams
 	{
 		return ("Width: " +width+", "+"hight: "+ hight+", resolution: "+resolution+", range_x"+ range_x.toString()+", range_y"+range_y.toString());
 	}
-	public double getWidth() {
+	public int getWidth() {
 		return width;
 	}
-	public void setWidth(double width) {
+	public void setWidth(int width) {
 		this.width = width;
 	}
-	public double getHight() {
+	public int getHight() {
 		return hight;
 	}
-	public void setHight(double hight) {
+	public void setHight(int hight) {
 		this.hight = hight;
 	}
-	public double getResolution() {
+	public int getResolution() {
 		return resolution;
 	}
-	public void setResolution(double resolution) {
+	public void setResolution(int resolution) {
 		this.resolution = resolution;
 	}
 	public Range getRange_x() {
@@ -49,4 +49,5 @@ public class GuiParams
 	public void setRange_y(Range range_y) {
 		this.range_y = range_y;
 	}
+	
 }
